refactoring to avoid circular import when importing torch models
